environment_nlp_gradio_separated/requirements.txt
gradio
transformers
torch
diffusers
matplotlib
accelerate
safetensors
Pillow